Seminars/Seminar_6/ml_pipeline/preprocess/requirements.txt
pillow==11.3.0
numpy==2.2.6
fastapi==0.120.2
requests==2.32.5
python-multipart==0.0.20
uvicorn==0.38.0